backend/requirements.txt
fastapi[all]==0.115.12
SQLAlchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
neo4j==5.28.1
redis==5.2.1
python-dotenv==1.0.1
